Combine 2 files related to Ad Blocking Recovery into one.
