+++ assets/js/modules/adsense/datastore/index.js
@@ -33,7 +33,6 @@ import adblocker from './adblocker';
 import service from './service';
 import sites from './sites';
 import adBlockingRecovery from './ad-blocking-recovery';
-import adBlockingRecoveryTags from './ad-blocking-recovery-tags';
 import { MODULES_ADSENSE } from './constants';
 
 const store = Data.combineStores(
@@ -49,8 +48,7 @@ const store = Data.combineStores(
 	adblocker,
 	service,
 	sites,
-	adBlockingRecovery,
-	adBlockingRecoveryTags
+	adBlockingRecovery
 );
 
 export const initialState = store.initialState;
